requirements.txt
python-telegram-bot==11.1.0
python-dotenv==0.12.0
requests==2.23.0
redis==3.4.1
validate-email==1.3
